requirements.txt
aiogram>=3
aiosqlite
APScheduler
pydantic
python-dotenv
matplotlib
openpyxl
